requirements.txt
777==0.0.1
absl-py==2.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp_socks==0.9.0
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.35.24
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
boto3==1.35.57
botocore==1.35.58
bottle==0.12.25
Brotli==1.1.0
browser-cookie3==0.19.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
chmod==0.0.8
click==8.1.7
cloudscraper==1.2.71
clr-loader==0.2.6
colorama==0.4.6
contourpy==1.3.0
cookiecutter==2.6.0
cryptography==43.0.0
cssselect2==0.7.0
curl_cffi==0.7.1
cycler==0.12.1
Cython==3.0.11
decorator==4.4.2
defusedxml==0.7.1
Django==5.0.7
docker==7.1.0
docutils==0.16
duckduckgo_search==6.2.10
easyocr==1.7.2
et-xmlfile==1.1.0
fastapi==0.112.1
filelock==3.16.1
filetype==1.2.0
fire==0.7.0
Flask==3.0.3
flet==0.23.2
flet-core==0.23.2
flet-runtime==0.23.2
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.10.0
future==1.0.0
g4f==0.3.3.4
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.1
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hub-sdk==0.0.15
huggingface-hub==0.26.2
idna==3.7
imageio==2.35.0
imageio-ffmpeg==0.5.1
iso8601==2.1.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
Jinja2==3.1.4
jmespath==1.0.1
keyboard==0.13.5
keyring==8.7
keyrings.alt==5.0.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.8
loguru==0.7.2
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.4.2
ngrok==1.4.0
ninja==1.11.1.1
Nuitka==2.4.7
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
Opy==1.1.28
ordered-set==4.1.0
packaging==23.2
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
plyer==2.1.0
primp==0.6.1
proglog==0.1.10
protobuf==5.28.3
proxy_tools==0.1.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyarmor==8.5.11
pyarmor.cli.core==6.5.3
pyasn1==0.6.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pybboxes==0.1.6
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyExecJS==1.5.1
pygame==2.5.2
pygbag==0.9.0
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyMsgBox==1.0.9
pyngrok==7.2.0
pynput==1.7.7
pyparsing==3.2.0
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
Pyro4==4.82
PyScreeze==0.1.30
pyserial==3.5
pyTelegramBotAPI==4.21.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.5.1
python-telegram-bot==21.4
pythonnet==3.0.3
pytorch-lightning==2.4.0
pytweening==1.2.0
pytz==2024.1
pywebview==5.2
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.49
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.3
safetensors==0.4.5
sahi==0.11.18
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
serial==0.0.97
serpent==1.41
setuptools==69.2.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spark-parser==1.8.9
sqlparse==0.5.1
srt==3.5.3
starlette==0.38.2
sympy==1.13.1
tabulate==0.9.0
telebot==0.0.5
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
text-unidecode==1.3
thop==0.1.1.post2209072238
threaded==4.2.0
tifffile==2024.9.20
timm==1.0.11
tinycss2==1.3.0
torch==2.5.1
torchmetrics==1.5.2
torchvision==0.20.1
tqdm==4.66.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.28
ultralytics-thop==2.0.11
uncompyle6==3.9.2
urllib3==2.2.2
uvicorn==0.30.6
vidstream==0.0.14
voluptuous==0.15.2
vosk==0.3.45
watchdog==4.0.2
watchfiles==0.23.0
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
win32-setctime==1.1.0
xdis==6.1.1
yarl==1.9.4
yolo==0.3.1
yolov5==7.0.13
yolov8==0.0.2
zstandard==0.23.0
